+++ changelog.md
@@ -10,6 +10,7 @@
 version 4.7.0 - upcomming
   * included [cassandra plugin](/milkman-cassandra) and [Grpc plugin](/milkman-grpc) into macos app bundle
   * reordered toolbar-actions
+  * added image-preview for responses
   
 version 4.6.0
   * added option to disable ui animations